build deploy script

+++ provisioners/windows/MDC1Windows/OS-deploy.ps1
@@ -108,7 +108,12 @@ $OS_files = $local_install + $neededImage
 $setup = $local_install + $neededImage + "\setup.exe"
 
 if ((!(Test-Path $setup)) {
-    Remove-Item -Path "$local_install:\*" -Recurse -Force
+    Write-Host Install files wrong or missing
+    Write-Host Will resync
+    if ((Get-ChildItem -Path $local_install -Force).Count -gt 0) {
+        Write-Host Wrong install files - REMOVING
+        Remove-Item -Path "$local_install:\*" -Recurse -Force
+    }
     ## Mount Deployment share
     ## PSDrive is will unmount when the Powershell sessions ends. Ultimately maybe OK.
     ## net use will presist
